+++ Packages/dev.unity2foxglove.sdk/Runtime/Schemas/Proto/Publishers/FoxgloveCameraPublisher.Video.cs
@@ -219,7 +219,12 @@ namespace Unity.FoxgloveSDK.Components
                 _pendingRequests,
                 out var message);
             if (message != null)
-                Debug.Log(message);
+                Debug.LogFormat(
+                    LogType.Log,
+                    LogOption.NoStacktrace,
+                    this,
+                    "{0}",
+                    message);
         }
 
         /// <summary>